+++ javaScript/projects/weather-app/src/components/tabs/forecast/forecast.config.js
@@ -1,5 +1,6 @@
 // forecastday.date |
-// (forecastday.day.maxtemp_f / forecastday.day.mintemp_f)
+// (forecastday.day.maxtemp_f
+// forecastday.day.mintemp_f)
 // forecastday.day.condition
 // forecastday.day.daily_chance_of_rain |
 // wind
@@ -7,28 +8,136 @@
 // ENE, E, ESE
 // SE, SSE, S, SSW, SW
 // WSW, W, WNW
+import importAll from '../../../helpers/importAll';
+import formatTime from '../../../helpers/formatTime';
+import formatDate from '../../../helpers/formatDate';
+
+// const icons = importAll(require.context('../../../assets/icons', false, /\.svg$/));
 const unitSystems = {
-  imperial: [],
-  metric: [],
+  icons: importAll(require.context('../../../assets/icons', false, /\.svg$/)),
+  metric: {
+    temp: 'c',
+    speed: 'kph',
+    precipitation: 'mm',
+    pressure: 'mb',
+    distance: 'km',
+  },
+  imperial: {
+    temp: 'f',
+    speed: 'mph',
+    precipitation: 'in',
+    pressure: 'in',
+    distance: 'miles',
+  },
+  get(key) {
+    return this.unitSystem[key];
+  },
+  setIcon(key) {
+    return this.icons.files[Object.keys(this.icons.files).find((iconKey) => iconKey.includes(key))];
+  },
+  roundValue(value) {
+    return Math.round(value);
+  },
+  setValue(obj, ...args) {
+    return this.roundValue(obj[`${args[0]}${this.get(args[1])}`]);
+  },
 };
 
-const forecast = {
-  current: {
-    last_updated: null,
+const data = (state) => [
+  {
+    key: 'name',
+    value: state.date,
+    setText() {
+      return `${formatDate(this.value)}`;
+    },
   },
+  {
+    key: 'minmaxtemp',
+    min: {
+      value: state.setValue(state, 'mintemp_', 'temp'),
+      label: 'low',
+    },
+    max: {
+      value: state.setValue(state, 'maxtemp_', 'temp'),
+      label: 'high',
+    },
+    unit: '°',
+    label: `high / low`,
+    setText() {
+      return `${this.max.value}${this.unit} / ${this.min.value}${this.unit}`;
+    },
+  },
+  {
+    key: 'condition',
+    label: state.condition.text,
+    icon: state.condition.icon,
+    setText() {
+      return `${this.label}`;
+    },
+  },
+  {
+    key: 'precip',
+    value: state.daily_chance_of_rain,
+    unit: '%',
+    icon: state.setIcon('rain'),
+    setText() {
+      return `${this.value}${this.unit}`;
+    },
+  },
+];
+
+const location = (state) => ({
   location: {
-    name: null,
+    country: state.country,
+    localtime: state.localtime,
+    name: state.name,
+    region: state.region,
+    setText() {
+      return `${this.name}, ${
+        this.region.length === 0 || this.region === this.name ? this.country : this.region
+      }`;
+    },
   },
-  forecastday: [],
-};
+});
 
 const forecastController = {
-  init(unitSystem, weatherData) {},
-  setValues(key, subKey, item) {},
-  findObjects(key) {},
+  init(weatherData, unitSystem) {
+    this.weatherData = weatherData;
+    this.unitSystem = unitSystem;
+    this.setDay = this.setDay.bind(this);
+    const forecastday = weatherData.forecast.forecastday.map(this.setDay);
+    const state = {
+      ...weatherData.location,
+    };
+
+    return {
+      ...location(state),
+      forecastday,
+      last_updated: formatTime(weatherData.current.last_updated).toLowerCase(),
+    };
+  },
+  setDay(obj) {
+    const days = obj;
+    const state = {
+      icons: unitSystems.icons,
+      get: unitSystems.get,
+      setIcon: unitSystems.setIcon,
+      setValue: unitSystems.setValue,
+      roundValue: unitSystems.roundValue,
+      unitSystem: unitSystems[this.unitSystem],
+      ...obj.day,
+      ...obj,
+    };
+
+    return { ...data(state) };
+  },
 };
 
 export default {
-  init(unitSystem, weatherData) {},
-  setProperties() {},
+  init(weatherData, unitSystem, timeStamp) {
+    this.setProperties(forecastController.init(weatherData, unitSystem));
+  },
+  setProperties(obj) {
+    Object.assign(this, obj);
+  },
 };