+++ javaScript/projects/weather-app/src/components/tabs/forecast/forecast.js
@@ -1,6 +1,7 @@
 import createElement from '../../../helpers/createElement';
 import forecast from './forecast.config';
 import createContentRows from '../../../helpers/createContentRows';
+import formatTime from '../../../helpers/formatTime';
 
 const forecastBuilder = {
   init() {},
@@ -12,17 +13,42 @@ const forecastBuilder = {
   },
   render() {
     console.log('render() running from forecast.js');
+    console.log(forecast);
     const forecastSection = createElement('section');
-    const forecastSectionHeading = createElement('h1');
-    forecastSection.id = 'forecast';
-    forecastSectionHeading.setAttributes({ textContent: '3 Day Weather' });
-    forecastSection.appendChild(forecastSectionHeading);
+    const forecastSectionHeader = createElement('header');
+    const forecastSectionHeading = createElement('h2');
+    const forecastLocation = createElement('span');
+    const forecastTimeStamp = createElement('p');
 
+    forecastSection.id = 'forecast';
+    forecastSectionHeading.textContent = '3 Day Weather';
+    forecastLocation.textContent = ` - ${forecast.location.setText()}`;
+    forecastTimeStamp.textContent = `As of ${forecast.last_updated}`;
+
+    forecastSectionHeading.appendChild(forecastLocation);
+    forecastSectionHeader.appendChild(forecastSectionHeading);
+    forecastSectionHeader.appendChild(forecastTimeStamp);
+    forecastSection.appendChild(forecastSectionHeader);
+
+    // temporary
+    const forecastDetails = createElement('section');
+    forecastDetails.id = 'forecast_details';
+
+    forecast.forecastday.forEach((day) => {
+      Object.values(day).forEach((detail) => {
+        forecastDetails.append(
+          createContentRows(createElement, null, detail.icon, detail.setText()),
+        );
+      });
+    });
+    // temporary
+    forecastSection.appendChild(forecastDetails);
     return forecastSection;
   },
 };
 
-export default function buildForecast(weatherData) {
+export default function buildForecast(weatherData, timeStamp) {
+  forecast.init(weatherData, 'imperial', timeStamp);
   return forecastBuilder.render();
 }