+++ javaScript/projects/weather-app/src/components/tabs/today/today.config.js
@@ -7,10 +7,11 @@
 // current.vis_miles
 // current.wind_mph
 // current.wind_dir
+import formatTime from '../../../helpers/formatTime';
 import importAll from '../../../helpers/importAll';
 
-const icons = importAll(require.context('../../../assets/icons', false, /\.svg$/));
 const unitSystems = {
+  icons: importAll(require.context('../../../assets/icons', false, /\.svg$/)),
   metric: {
     temp: 'c',
     speed: 'kph',
@@ -29,7 +30,7 @@ const unitSystems = {
     return this.unitSystem[key];
   },
   setIcon(key) {
-    return icons.files[Object.keys(icons.files).find((iconKey) => iconKey.includes(key))];
+    return this.icons.files[Object.keys(this.icons.files).find((iconKey) => iconKey.includes(key))];
   },
   roundValue(value) {
     return Math.round(value);
@@ -40,38 +41,42 @@ const unitSystems = {
 };
 
 const data = (state) => ({
-  summary: {
-    condition: {
-      label: state.current.condition.text,
-      icon: state.current.condition.icon,
-      setText() {
-        return `${this.label}`;
-      },
-    },
-    temp: {
+  summary: [
+    {
+      key: 'temp',
       unit: '°',
-      value: state.setValue(state.current, 'temp_', 'temp'),
+      value: state.setValue(state, 'temp_', 'temp'),
       setText() {
         return `${this.value}${this.unit}`;
       },
     },
-    feelslike: {
+    {
+      key: 'condition',
+      label: state.condition.text,
+      icon: state.condition.icon,
+      setText() {
+        return `${this.label}`;
+      },
+    },
+  ],
+  details: [
+    {
+      key: 'feelslike',
       unit: '°',
-      value: state.setValue(state.current, 'feelslike_', 'temp'),
+      value: state.setValue(state, 'feelslike_', 'temp'),
       label: 'feels like',
       setText() {
         return `${this.value}${this.unit}`;
       },
     },
-  },
-  details: {
-    minmaxtemp: {
+    {
+      key: 'minmaxtemp',
       min: {
-        value: state.setValue(state.forecast.forecastday[0].day, 'mintemp_', 'temp'),
+        value: state.setValue(state.day, 'mintemp_', 'temp'),
         label: 'low',
       },
       max: {
-        value: state.setValue(state.forecast.forecastday[0].day, 'maxtemp_', 'temp'),
+        value: state.setValue(state.day, 'maxtemp_', 'temp'),
         label: 'high',
       },
       unit: '°',
@@ -81,8 +86,9 @@ const data = (state) => ({
         return `${this.max.value}${this.unit} / ${this.min.value}${this.unit}`;
       },
     },
-    humidity: {
-      value: state.current.humidity,
+    {
+      key: 'humidity',
+      value: state.humidity,
       unit: '%',
       label: 'humidity',
       icon: state.setIcon('humidity'),
@@ -90,8 +96,9 @@ const data = (state) => ({
         return `${this.value}${this.unit}`;
       },
     },
-    pressure: {
-      value: state.setValue(state.current, 'pressure_', 'pressure'),
+    {
+      key: 'pressure',
+      value: state.setValue(state, 'pressure_', 'pressure'),
       unit: state.get('pressure'),
       label: 'pressure',
       icon: state.setIcon('pressure'),
@@ -99,8 +106,9 @@ const data = (state) => ({
         return `${this.value} ${this.unit}`;
       },
     },
-    vis: {
-      value: state.setValue(state.current, 'vis_', 'distance'),
+    {
+      key: 'vis',
+      value: state.setValue(state, 'vis_', 'distance'),
       unit: state.get('distance'),
       label: 'visibility',
       icon: state.setIcon('visibility'),
@@ -108,26 +116,40 @@ const data = (state) => ({
         return `${this.value} ${this.unit}`;
       },
     },
-    wind: {
-      value: state.setValue(state.current, 'wind_', 'speed'),
+    {
+      key: 'wind',
+      value: state.setValue(state, 'wind_', 'speed'),
       unit: state.get('speed'),
       label: 'wind',
       icon: state.setIcon('wind'),
       dir: {
-        value: state.current.wind_dir,
+        value: state.wind_dir,
       },
       setText() {
         return `${this.dir.value} ${this.value} ${this.unit}`;
       },
     },
-    uv: {
-      value: state.current.uv,
+    {
+      key: 'uv',
+      value: state.uv,
       label: 'UV Index',
       icon: state.setIcon('uv'),
       setText() {
         return `${this.value} of 11`;
       },
     },
+  ],
+});
+
+const location = (state) => ({
+  country: state.country,
+  localtime: state.localtime,
+  name: state.name,
+  region: state.region,
+  setText() {
+    return `${this.name}, ${
+      this.region.length === 0 || this.region === this.name ? this.country : this.region
+    }`;
   },
 });
 
@@ -136,14 +158,21 @@ const todayController = {
     this.weatherData = weatherData;
     this.unitSystem = unitSystem;
     const state = {
+      icons: unitSystems.icons,
       get: unitSystems.get,
       setIcon: unitSystems.setIcon,
       setValue: unitSystems.setValue,
       roundValue: unitSystems.roundValue,
       unitSystem: unitSystems[unitSystem],
-      ...weatherData,
+      ...weatherData.current,
+      ...weatherData.forecast.forecastday[0],
+      ...weatherData.location,
     };
-    return { ...data(state) };
+    return {
+      ...data(state),
+      location: location(state),
+      last_updated: formatTime(weatherData.current.last_updated).toLowerCase(),
+    };
   },
 };